Move Vale next to the other grammar checkers under writer/locale
Vale is a grammar/style provider like Harper and LanguageTool, not a
scripting-venv helper. Keep git history with git mv, update imports and
LibrePy excludes, and leave the host RPC facade in scripting.client.
