+++ docs/archive/scripting-domain-debt-dev-plan.md
@@ -65,7 +65,7 @@ Follow-ons from the Phase 6 planning pass. Prefer least-code reuse of existing f
 |------|--------|
 | Extract result formatting from [`python_runner.py`](../plugin/scripting/python_runner.py) | **Skipped.** LibrePy Run Python Script still formats Writer HTML / Calc values after the venv returns; extracting to `result_formatters.py` is churn, not a quality win. |
 | Thin [`venv_diagnostics.py`](../plugin/scripting/venv_diagnostics.py) | Declarative probe-group specs (reuse `_self_check_group_specs` shape); keep orchestration thin. Heavy existing coverage makes this low risk. Do not slim this file for LibrePy while it still works. |
-| Targeted unit tests | Direct tests for [`trusted_rpc.py`](../plugin/scripting/trusted_rpc.py) are in [`tests/scripting/test_trusted_rpc.py`](../tests/scripting/test_trusted_rpc.py). Calc insert tests are in [`tests/scripting/test_python_runner_formatting.py`](../tests/scripting/test_python_runner_formatting.py). Draw insert stays WIP (msgbox + commented block). Optional [`venv/vale.py`](../plugin/scripting/venv/vale.py) unit coverage. |
+| Targeted unit tests | Direct tests for [`trusted_rpc.py`](../plugin/scripting/trusted_rpc.py) are in [`tests/scripting/test_trusted_rpc.py`](../tests/scripting/test_trusted_rpc.py). Calc insert tests are in [`tests/scripting/test_python_runner_formatting.py`](../tests/scripting/test_python_runner_formatting.py). Draw insert stays WIP (msgbox + commented block). Optional [`vale.py`](../plugin/writer/locale/vale.py) unit coverage. |
 | Harper path docs only | Host in-process grammar vs trusted-worker path is intentional: many users never configure a Python venv, and grammar must still work. Do not route realtime Harper through the warm worker. |
 
 ### Larger / higher risk — defer unless needed
