+++ plugin/scripting/sandbox.py
@@ -153,7 +153,7 @@ VENV_AUTHORIZED_IMPORTS: tuple[str, ...] = (
     "plugin.scripting.forecast",
     "plugin.scripting.calc_functions",
     "plugin.scripting.calc_functions.*",
-    "plugin.scripting.venv.vale",
+    "plugin.writer.locale.vale",
     "plugin.writer.locale.languagetool",
 )
